// src/app/audio-visual/model/audio-visual.model.ts
export class AudioVisualModel {

    _001: AudioVisualDetailsModel;
    _002: AudioVisualDetailsModel;
    _003: AudioVisualDetailsModel;
    _004: AudioVisualDetailsModel;
    _005: AudioVisualDetailsModel;
    _005a:AudioVisualDetailsModel;
    _006: AudioVisualDetailsModel;
    _007: AudioVisualDetailsModel;
    _008: AudioVisualDetailsModel;
    _009: AudioVisualDetailsModel;
    _010: AudioVisualDetailsModel;
    _011: AudioVisualDetailsModel;
    _012: AudioVisualDetailsModel;
    _014: AudioVisualDetailsModel;
    _013: AudioVisualDetailsModel;
    _015: AudioVisualDetailsModel;
    _016: AudioVisualDetailsModel;
    _017: AudioVisualDetailsModel;
    _018: AudioVisualDetailsModel;
    _019: AudioVisualDetailsModel;
    _020: AudioVisualDetailsModel;
    _021: AudioVisualDetailsModel;
    _022: AudioVisualDetailsModel;
    _023: AudioVisualDetailsModel;
    _024: AudioVisualDetailsModel;
    _025: AudioVisualDetailsModel;
    _026: AudioVisualDetailsModel;
    _027: AudioVisualDetailsModel;
    _028: AudioVisualDetailsModel;
    _029: AudioVisualDetailsModel;
    _030: AudioVisualDetailsModel;
    _031: AudioVisualDetailsModel;
    _032: AudioVisualDetailsModel;
    _033: AudioVisualDetailsModel;
    _034: AudioVisualDetailsModel;
    _035: AudioVisualDetailsModel;
    _036: AudioVisualDetailsModel;

    constructor() { 
		this._001  =  new AudioVisualDetailsModel();
		this._002  =  new AudioVisualDetailsModel();
		this._003  =  new AudioVisualDetailsModel();
		this._004  =  new AudioVisualDetailsModel();
		this._005  =  new AudioVisualDetailsModel();
		this._005a =  new AudioVisualDetailsModel();
		this._006  =  new AudioVisualDetailsModel();
		this._007  =  new AudioVisualDetailsModel();
		this._008  =  new AudioVisualDetailsModel();
		this._009  =  new AudioVisualDetailsModel();
		this._010  =  new AudioVisualDetailsModel();
		this._011  =  new AudioVisualDetailsModel();
		this._012  =  new AudioVisualDetailsModel();
		this._014  =  new AudioVisualDetailsModel();
		this._013  =  new AudioVisualDetailsModel();
		this._015  =  new AudioVisualDetailsModel();
		this._016  =  new AudioVisualDetailsModel();
		this._017  =  new AudioVisualDetailsModel();
		this._018  =  new AudioVisualDetailsModel();
		this._019  =  new AudioVisualDetailsModel();
		this._020  =  new AudioVisualDetailsModel();
		this._021  =  new AudioVisualDetailsModel();
		this._022  =  new AudioVisualDetailsModel();
		this._023  =  new AudioVisualDetailsModel();
		this._024  =  new AudioVisualDetailsModel();
		this._025  =  new AudioVisualDetailsModel();
		this._026  =  new AudioVisualDetailsModel();
		this._027  =  new AudioVisualDetailsModel();
		this._028  =  new AudioVisualDetailsModel();
		this._029  =  new AudioVisualDetailsModel();
		this._030  =  new AudioVisualDetailsModel();
        this._031  =  new AudioVisualDetailsModel();
        this._032  =  new AudioVisualDetailsModel();
        this._033  =  new AudioVisualDetailsModel();
        this._034  =  new AudioVisualDetailsModel();
        this._035  =  new AudioVisualDetailsModel();
        this._036  =  new AudioVisualDetailsModel();
	}

}

export class AudioVisualDetailsModel {

    slNo: string | number = '';
    checked: boolean = false;
    particulars: string = '';
    qty: string | number = '';
    rate: string = '';
    active: string = '';

    constructor() { }
    
}